Setup `rake test` to run the tests
This updates the Rakefile to allow basic running of the tests. However,
I found that running the entire test suite eats up all my memory (and
then some), and eventually produces errors in the Quickdraw Runner.
Because it eats up a bunch of memory, it makes it hard to troubleshoot
the Quickdraw error. However, the fact it consumes so much memory for
what appears to be a small test suite is probably a bigger issue. The
README for Quickdraw itself says it's not ready for production use, so I
figured I'd leave the memory issue alone for now.

Any additional arguments are passed on to the test runner, so you can do
things like `rake test -t 3 test/path/to/testfile.test.rb`.

Closes yippee-fun/plume#10